refactor(setbuilder): split 1400-line pass2_agent.py into focused modules
The agent-toolkit module had grown to 1401 lines (global cap: 800) and was the
epicenter of the recent Family 1/2 merge-conflict cascades — every new tool piled
into the same MUTATION_TOOLS set, apply_tool_call dispatcher, handler block,
display-summary switch, and _agent_tools() registry in one file.

Extract along the mutating-vs-read-only seam (which maps onto the existing
MUTATION_TOOLS set), leaving pass2_agent.py as a thin orchestration + re-export
facade so no call site changes:

  agent_common.py          errors, the mutation allowlist, shared DB lookups
  agent_tools_mutations.py 15 mutating handlers + helpers
  agent_tools_sensing.py   6 read-only handlers + helpers
  agent_tool_specs.py      the tool-spec registry
  agent_display.py         ToolCard one-line summaries
  pass2_agent.py           chat/critique orchestration + dispatcher (410 LOC)

Pure refactor, no behavior change: 2992 tests pass, coverage 88.47% (gate 85%),
bandit clean. Unblocks Family 3 (#442) landing in small new files instead of
growing the hotspot.
